피보나치 수

+++ Algorithm/src/Programmers/피보나치수.cpp
@@ -3,15 +3,15 @@
 
 using namespace std;
 
-long long solution(int n) {
-    long long answer = 0;
-    vector<long long> v;
+int solution(int n) {
+    int answer = 0;
+    vector<int> v;
     
     v.push_back(0);
     v.push_back(1);
 
     for(int i = 2; i <= n; i++){
-        long long tmp = v[i-2] + v[i-1];
+        int tmp = v[i-2] + v[i-1];
         if(tmp > 1234567) tmp %= 1234567;
         v.push_back(tmp);
     }
